python/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rtclient"
version = "0.5.3"
description = "A client for the RT API"
authors = ["ZhiPu Corporation"]

[tool.poetry.dependencies]
python = ">=3.10"
aiohttp = "*"
pydantic = "*"
python-dotenv = "^1.0.1"

[tool.poetry.extras]
audio = ["pyaudio"]
[tool.poetry.dev-dependencies]
ruff = "*"
black = "*"
python-dotenv = "*"

pytest = "*"
pytest-asyncio = "*"

[tool.poetry.scripts]


[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
extend-ignore = ["UP007"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
